# 20% faster `try`